[add] 1) add check rights

+++ controllers/EventController.php
@@ -3,16 +3,26 @@
 namespace app\controllers;
 
 use app\models\Event;
+use app\models\User;
 use app\models\EventSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use Yii;
 
 /**
  * EventController implements the CRUD actions for Event model.
  */
 class EventController extends Controller
 {
+    private ?User $user;
+
+    public function __construct($id, $module, $config = []) {
+        $this->user = Yii::$app->user->isGuest ? null : Yii::$app->user->getIdentity()->user;
+        parent::__construct($id, $module, $config);
+    }
     /**
      * @inheritDoc
      */
@@ -21,6 +31,15 @@ class EventController extends Controller
         return array_merge(
             parent::behaviors(),
             [
+                'access' => [
+                    'class' => AccessControl::class,
+                    'rules' => [
+                        [
+                            'allow' => true,
+                            'roles' => ['@'],
+                        ],
+                    ],
+                ],
                 'verbs' => [
                     'class' => VerbFilter::class,
                     'actions' => [
@@ -69,6 +88,9 @@ class EventController extends Controller
     {
         $model = new Event();
         $model->project_id = $projectId;
+        if(!$model->project->canEdit($this->user)) {
+            throw new ForbiddenHttpException('Вы не можете создавать мероприятие на этот проект!');
+        }
 
         if ($this->request->isPost) {
             if ($model->load($this->request->post()) && $model->save()) {
@@ -93,6 +115,9 @@ class EventController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
+        if(!$model->project->canEdit($this->user)) {
+            throw new ForbiddenHttpException('Вы не можете изменять мероприятие этого проекта!');
+        }
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
             return $this->redirect(['project/view', 'id' => $model->project_id, 'tab' => 'event']);
@@ -113,6 +138,9 @@ class EventController extends Controller
     public function actionDelete($id)
     {
         $model = $this->findModel($id);
+        if(!$model->project->canEdit($this->user)) {
+            throw new ForbiddenHttpException('Вы не можете удалять мероприятие этого проекта!');
+        }
         $model->delete();
 
         return $this->redirect(['project/view', 'id' => $model->project_id, 'tab' => 'event']);
